phase-12-genai/requirements.txt
anthropic>=0.40.0
psycopg2-binary>=2.9
prometheus-client>=0.20.0
pydantic>=2.0
apache-airflow>=2.8.0
google-cloud-bigquery>=3.0.0
